Move API routes to separate module
